% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{calc_filtration_density}
\alias{calc_filtration_density}
\title{Calculate the age filtration density}
\usage{
calc_filtration_density(xcalc, x_mid, infant_prop, discrete = T)
}
\arguments{
\item{xcalc}{The vector of ages for calculation}

\item{x_mid}{The midpoint age for the filtration density}

\item{infant_prop}{The proportion of infants in the filtration density}

\item{discrete}{Whether to return a discrete density [default: TRUE]}
}
\value{
The age filtration density
}
\description{
Calculate the age filtration density
}
