% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{calc_trapez_weights}
\alias{calc_trapez_weights}
\title{Calculate trapezoidal integration weights}
\usage{
calc_trapez_weights(tau)
}
\arguments{
\item{tau}{A vector of grid points.}
}
\value{
A vector of trapezoidal integration weights.
}
\description{
This function calculates the weights for trapezoidal integration given a
vector of grid points.
}
\details{
The function computes the weights for trapezoidal integration
  based on the provided grid points. The weights for the central points
  are calculated as the average of the differences between the adjacent
  grid points. The weights for the first and last points are calculated
  as half the difference between the first two and last two grid points,
  respectively.
}
\examples{
tau <- seq(0, 1, by = 0.1)
weights <- calc_trapez_weights(tau)

}
