% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{check_siler_inputs}
\alias{check_siler_inputs}
\title{A utility function to check Siler inputs (b and x0)}
\usage{
check_siler_inputs(b, x0)
}
\arguments{
\item{b}{The parameter vector}

\item{x0}{The conditional starting age}
}
\value{
None. An error is thrown if the inputs fail the checks.
}
\description{
Ensure that b is positive. Ensure that x0>=0.
}
