% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{chsiler}
\alias{chsiler}
\title{Calculate the Siler cumulative hazard}
\usage{
chsiler(x, a, x0 = 0)
}
\arguments{
\item{x}{The parameter vector of ages}

\item{a}{The parameter vector}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The vector of cumulative hazards (same length as x)
}
\description{
Calculate the Siler cumulative hazard
}
