% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{demohaz_to_trad_siler_param}
\alias{demohaz_to_trad_siler_param}
\title{Convert from the demohaz to traditional Siler parameterization}
\usage{
demohaz_to_trad_siler_param(b)
}
\arguments{
\item{b}{The demohaz pramaterization of the Siler hazard}
}
\value{
The traditional parameterization of the Siler hazard
}
\description{
For the input vector a in the traditional parameterization, convert to the
demohaz parameterization. The traditional parameterization of the hazard is
a[1]*exp(a[2]*x) + a[3] + a[4]*exp(a[5]*x) whereas the demohaz
parameterization is b[1]*exp(b[2]*x) + b[3] + b[5]*exp(b[5]*(x-b[4])). They
are related via

a[1] = b[1]
a[2] = b[2]
a[3] = b[3]
a[4] = b[5]*exp(-b[4]*b[5])
a[5] = b[5]
}
