% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{do_mh_sampling_at_temp}
\alias{do_mh_sampling_at_temp}
\title{Do Metropolis-Hastings sampling at a given temperature}
\usage{
do_mh_sampling_at_temp(
  init,
  neg_log_cost_func = NA,
  num_samp = NA,
  temp = NA,
  prop_scale = NA,
  save_theta = NA,
  ...
)
}
\arguments{
\item{init}{The initializing object (either a parameter vector or a chain
created by a previous call to do_mh_sampling_at_temp).}

\item{neg_log_cost_func}{The function that calculates the negative log of the
function being sampled.}

\item{num_samp}{The number of samples to make.}

\item{temp}{The temperature}

\item{prop_scale}{The standard deviation(s) to use for the proposal
distribution (either a vector that is the same length as the parameter
vector, theta, or a scalar that is used for all parameters).}

\item{save_theta}{Whether or not to save and return the accepted parameter
values (assumed False if not input). Cannot be over-ridden for continuing
chains.}

\item{...}{Variables required by neg_log_cost_func.}
}
\value{
A list-like object of class "mh_chain" with sampling information
}
\description{
Do Metropolis-Hastings sampling given an input negative log of some target
distribution (probably a negative log-likelihood) at a given temperature. The
chain can be initialized either with a starting parameter vector or with a
chain created by a previous call to do_mh_sampling_at_temp. For continuing
chains, some (but not all) of the inputs from the previously used values can
be input.
}
