% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{fast_gradnllsiler}
\alias{fast_gradnllsiler}
\title{Fast calculation of the gradient of the Siler negative log-likelihood}
\usage{
fast_gradnllsiler(b, xvalues, xcounts, x0 = 0)
}
\arguments{
\item{b}{Siler parameter vector}

\item{xvalues}{x values for the calculation}

\item{xcounts}{The number of observations for each entry in xvalues}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The gradient of the negative log-likelihood (a vector of length 5)
}
\description{
Calculate the gradient of the Siler negative log-likelihood using tabulated
values (xvalues and xcounts; see the documentation for fast_nllsiler).
}
