% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{fast_nllsiler}
\alias{fast_nllsiler}
\title{Use counts to quickly calculate the Siler negative log-likelihood}
\usage{
fast_nllsiler(b, xvalues, xcounts, x0 = 0)
}
\arguments{
\item{b}{The Siler parameter vector}

\item{xvalues}{Locations at which to evaluate probability density function}

\item{xcounts}{The number of observations for each entry in xvalues}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The negative log-likelihood
}
\description{
Calculate the negative log-likelihood of the Siler probability density.
This can often be done much faster by tabulating repeated x-values. If there
are no (or few) repeats, there is little speed loss from using tabulated
values. The inputs are a vector of x values, xvalues, and corresponding
counts, xcounts, both of which are pre-computed. These two vectors can be
created from the vector of ages, x, using the following code:

xtable <- table(x)
xvalues <- as.numeric(names(xtable))
xcounts <- as.numeric(xtable)

The Siler hazard is:

lambda(x) = b1 * exp(-b_2*x) + b_3 + b5 * exp(b5*(x-b4))
}
