% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{fast_transformed_nllsiler}
\alias{fast_transformed_nllsiler}
\title{Calculate the negative log-likelihood for the Siler distribution using counts
(hence fast) and assuming a transformed parameter vector.}
\usage{
fast_transformed_nllsiler(bbar, xvalues, xcounts, beta, x0 = 0, xmax = 120)
}
\arguments{
\item{bbar}{The transformed parameter vector, used to increase robustness of
fitting}

\item{xvalues}{Locations at which to evaluate probability density function}

\item{xcounts}{The number of observations for each entry in xvalues}

\item{beta}{The baseline parameter vector}

\item{x0}{The conditional starting age [default: 0].}

\item{xmax}{The baseline maximum age (default: 120)}
}
\value{
The negative log-likelihood
}
\description{
The vector of ages may contain repeats. If there are many repeats, the
negative log-likelihood calculation will be slow if it is repeated for each
repeated value in x. To avoid this, the inputs to this function are a
vector of xvalues and corresponding counts. These two vectors can be created
from the vector of ages, x, using the following code:

xtable <- table(x)
xvalues <- as.numeric(names(xtable))
xcounts <- as.numeric(xtable)

To improve the robustness of the negative log-likelihood fit, a transformed
vector is used for the actual optimization and the fourth parameter is
adjusted so that it is measured with respect to a notional maximum age,
xmax. To make this precise, define the vector b to be:

b1 = a1
b2 = a2
b3 = a3
b4 = a4 * exp(a5 * xmax)
b5 = a5

The Siler hazard given these definitions is:

lambda(x) = b1 * exp(-b_2*x) + b_3 + b4 * exp(b5*(x-xmax))

In addition, let beta be a baseline parameter vector and define the
transformed parameter vector bbar to be

b1 = beta1 * exp(bbar1)
b2 = beta2 * exp(bbar2)
b3 = beta3 * exp(bbar3)
b4 = beta4 * exp(bbar4)
b5 = beta5 * exp(bbar5)

This second transformation allows the parameter vector bbar to be
unconstrained (positive or negative) while the parameter vector b is
positive (assuming that the baseline vector beta is positive, which should
be ensured in its construction). In principle, the choices of beta and xmax
are arbitrary as far as the maximum likelihood optimization is concerned.
However, appropriate choices can make the fitting far more robust. If
fitting human demographic data, we suggest that xmax = 120 (years) and
derive beta from beta from Gage and Dyke 1986, Table 2, Level 15.
}
