% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{grad_chsiler_vect}
\alias{grad_chsiler_vect}
\title{Vectorized calculation of the gradient of the Siler cumulative hazard}
\usage{
grad_chsiler_vect(x, b, x0 = 0)
}
\arguments{
\item{x}{Vector of ages}

\item{b}{Siler parameter vector}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
A matrix of gradients with dimensions [length(x) x 5]
}
\description{
Calculate the gradient of the Siler cumulative hazard for each age in the
input vector x. This is a vectorized calculation since no sum is done over
x. Hence, the return object is a matrix with dimensions [N x 5] rather than
a vector of length 5 (N = length(x)). This allows fast calculation of the
gradient when there are repeated entries in x, as in gradsiler_fast.
}
