% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{gradnllsiler}
\alias{gradnllsiler}
\title{Calculate the gradient for the Siler distribution}
\usage{
gradnllsiler(b, x, x0 = 0)
}
\arguments{
\item{b}{The parameter vector}

\item{x}{The vector of ages}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The gradient vector (length 5)
}
\description{
Given the parameter vector b and the vector of ages x, calculate the gradient
for a negative log-likelihood fit. Do this symbolically using the Deriv
package.

See details in the documentation for hsiler for a definition of the Siler
hazard and related quantities.
}
