% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{hsiler_vect}
\alias{hsiler_vect}
\title{Vectorized calculation of the Siler hazard}
\usage{
hsiler_vect(x, b)
}
\arguments{
\item{x}{Vector of ages}

\item{b}{Siler parameter vector}
}
\value{
A matrix of hazards with dimensions [length(x) x 5]
}
\description{
Calculate the the Siler hazard, lambda, for each age in the input vector x.
This is a vectorized calculation that is used by, e.g., fast_gradnllsiler.
The return object is a matrix with dimensions [N x 5] rather than a vector
of length N (N = length(x)). The hazard is replicated 5 times for each
column to achieve these dimensions.
}
