% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load_demohaz_data}
\alias{load_demohaz_data}
\title{Load datasets in the demohaz R package}
\usage{
load_demohaz_data(dataset_name, raw = FALSE)
}
\arguments{
\item{dataset_name}{Name of the dataset to load}

\item{lifetable}{Whether to return the raw data (a dataframe) or the
lifetable}
}
\value{
The requested dataset
}
\description{
Load datasets in the demohaz R package. This is done with a function to
support reformating of the raw data text files and to make the associated
tests clearer. The following datasets can be loaded:
}
\details{
Ache       The Ache dataframe or lifetable
Hadza      The Hadza dataframe or lifetable
Hiwi       The Hiwi dataframe or lifetable
Tsimane    The Tsimane dataframe or lifetable
}
