% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{nll_usher3}
\alias{nll_usher3}
\title{Calculate the negative log-likelihood for the Usher 3 illness-death model}
\usage{
nll_usher3(theta, x, ill, x0 = 0)
}
\arguments{
\item{theta}{The parameter vector for the Usher illness-death model with
ordering [k1, k2, mort]. The mortality parameter vector, mort, is
[b1, b2, b3, b4, b5].}

\item{x}{The vector of ages-at-death}

\item{ill}{The vector of illness indicators, which can have NA entries}

\item{x0}{Conditional starting age [default: 0]}
}
\value{
The negative log-likelihood value
}
\description{
Calculate the negative log-likelihood for the Usher 3 illness-death model
}
