% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{nll_usher3_hessian_wrapper}
\alias{nll_usher3_hessian_wrapper}
\title{Wrapper function for the negative log-likelihood of the Usher 3 model}
\usage{
nll_usher3_hessian_wrapper(paramVect, ageVect, illVect, x0 = 0, x_cut = Inf)
}
\arguments{
\item{paramVect}{The parameter vector for the Usher illness-death model}

\item{ageVect}{The vector of ages-at-death}

\item{illVect}{The vector of illness indicators}

\item{x0}{Conditional starting age [default: 0]}

\item{x_cut}{The age at which the well-to-ill transition hazard becomes
zero [default: Inf]}
}
\value{
The negative log-likelihood value
}
\description{
Thin wrapper around \code{\link{nll_usher3}} that uses parameter names
(paramVect, ageVect, illVect) that do not collide with \code{numDeriv}'s
internal use of \code{x} as the differentiation variable.
}
\keyword{internal}
