% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{nll_usher3_log_hessian_wrapper}
\alias{nll_usher3_log_hessian_wrapper}
\title{Wrapper for Hessian/gradient of the log-transformed Usher 3 NLL}
\usage{
nll_usher3_log_hessian_wrapper(
  paramVect,
  ageVect,
  illVect,
  x0 = 0,
  x_cut = Inf
)
}
\arguments{
\item{paramVect}{The log-transformed parameter vector}

\item{ageVect}{The vector of ages-at-death}

\item{illVect}{The vector of illness indicators}

\item{x0}{Conditional starting age [default: 0]}

\item{x_cut}{The age at which the well-to-ill transition hazard becomes
zero [default: Inf]}
}
\value{
The negative log-likelihood value
}
\description{
Like nll_usher3_hessian_wrapper but exponentiates the parameter vector
first. Uses ageVect/illVect to avoid name collision with numDeriv's x
parameter.
}
\keyword{internal}
