% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{p_12_integrand}
\alias{p_12_integrand}
\title{Integrand for p_12 calculation}
\usage{
p_12_integrand(y, x0, k1, k2, b_siler)
}
\arguments{
\item{y}{The integration variable (age of illness onset)}

\item{x0}{The starting age}

\item{k1}{The constant transition rate from the healthy state to the ill
state}

\item{k2}{The factor by which the mortality hazard out of the ill state is
larger than that out of the healthy state}

\item{b_siler}{The parameter vector for the Siler hazard model}
}
\value{
The integrand value at y
}
\description{
Helper function for computing the integral in p_12(x0, x).
}
\details{
The integrand is: exp(-k1 * (y - x0)) * [S_13(x0, y)]^(1 - k2)
}
\keyword{internal}
