% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{q_2_integrand}
\alias{q_2_integrand}
\title{Integrand for q_2 calculation}
\usage{
q_2_integrand(y, x0, k1, k2, b_siler)
}
\arguments{
\item{y}{The integration variable (age of illness onset)}

\item{x0}{The starting age}

\item{k1}{The constant transition rate from the healthy state to the ill state}

\item{k2}{The factor by which the mortality hazard out of the ill state is
larger than that out of the healthy state}

\item{b_siler}{The parameter vector for the Siler hazard model}
}
\value{
The integrand value at y
}
\description{
Helper function for computing the integral in q_2(x0, x; w).
}
\keyword{internal}
