% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{qsiler}
\alias{qsiler}
\title{Calculate x given input quantile(s)}
\usage{
qsiler(qvect, a, x0 = 0)
}
\arguments{
\item{a}{The parameter vector}

\item{x0}{The conditional starting age [default: 0].}

\item{x}{The parameter vector of ages}
}
\value{
The vector of quantiles (same length as  x)
}
\description{
Solve the inverse problem for the cumulative density with a change of
variable
}
\details{
xbar = a5*(x - x0) / (a5*x + 1)

or, equivalently,

x = ((xbar/a5) + x0)/(1-xbar).

xbar lies between 0 to 1, which allows R's root-finder to be invoked on a
bounded interval. Normalizing by a5 regularizes the root finding for large
ages where the positive exponential term dominates the hazard.
}
