% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{rsiler}
\alias{rsiler}
\title{Make random samples from the Siler distribution}
\usage{
rsiler(N, a, x0 = 0)
}
\arguments{
\item{N}{The number of samples to make}

\item{a}{The parameter vector}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The vector of samples (length N)
}
\description{
Make random samples from the Siler distribution
}
