% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{temper_and_tune}
\alias{temper_and_tune}
\title{Solve an minimization problem by doing an initial tempering than fine-tuning
the optimization usking the Hooke-Jeeves algorithm.}
\usage{
temper_and_tune(
  obj_fun,
  th0,
  num_cyc = 100,
  samps_per_cyc = 20,
  temp_vect = 10^(rev(seq(-1, 1, by = 0.25))),
  prop_scale_mat = NULL,
  ...
)
}
\arguments{
\item{obj_fun}{The objective function to minimize}

\item{th0}{The starting point for minimization}

\item{num_cyc}{Number of cycles for the tempering (default: 100)}

\item{samps_per_cyc}{Samples percycle for the tempering (default: 20)}

\item{temp_vect}{Temperature vector for the tempering (default:
10^(rev(seq(-1,1,by=.25)))}

\item{prop_scale_mat}{Proposal scale matrix for the tempering (default:
ranges evenly from 0.1 for the hottest temperature to 0.001 for the
coldest temperature, for all parameter values at each temperature).}

\item{...}{Additional inputs to the objective function}
}
\value{
The best fit parameter vector
}
\description{
Many of the inputs are hard coded....
}
