% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{temper_usher3}
\alias{temper_usher3}
\title{Use parallel tempering to fit the usher3 model.}
\usage{
temper_usher3(
  th0 = c(0.01, 1, 0.175, 1.4, 0.368 * 0.01, log(0.917 * 0.1/(0.075 * 0.001))/(0.917 *
    0.1), 0.917 * 0.1),
  verbose = FALSE,
  fn_plot = NULL,
  num_cyc = 100,
  samps_per_cyc = 20,
  temp_vect = 10^(rev(seq(-1, 1, by = 0.25))),
  prop_scale_mat = NULL,
  lr = 1e-05,
  func_tol = 1e-06,
  miniter = 1,
  maxiter = 1000,
  report_period = 50,
  ...
)
}
\arguments{
\item{th0}{The initial parameter vector with the ordering [k1, k2, b_siler]
where b_siler uses the demohaz parameterization of the Siler hazard (see
Siler documentation)}

\item{verbose}{Whether to print verbose output [default: FALSE]}

\item{fn_plot}{The function for plotting [default: NULL]}

\item{num_cyc}{The number of cycles for tempering [default: 100]}

\item{samps_per_cyc}{The number of samples per cycle [default: 20]}

\item{temp_vect}{The temperature vector for tempering
[default: 10^(rev(seq(-1, 1, by = 0.25)))]}

\item{prop_scale_mat}{The proposal scaling matrix [default: NULL]}

\item{lr}{The learning rate for gradient descent [default: 1e-5]}

\item{func_tol}{The tolerance for the objective function [default: 1e-6]}

\item{miniter}{The minimum number of iterations [default: 1]}

\item{maxiter}{The maximum number of iterations [default: 1000]}

\item{report_period}{The reporting period for verbose output [default: 50]}

\item{...}{Additional arguments passed to obj_fun}
}
\value{
A list containing the results of tempering and tuning
}
\description{
Use parallel tempering to fit the usher3 model.
}
