% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{trad_to_demohaz_siler_param}
\alias{trad_to_demohaz_siler_param}
\title{Convert from the traditional to demohaz Siler parameterization}
\usage{
trad_to_demohaz_siler_param(a)
}
\arguments{
\item{a}{The traditional pramaterization of the Siler hazard}
}
\value{
The demohaz pramaterization of the Siler hazard
}
\description{
For the input vector a in the traditional parameterization, convert to the
demohaz parameterization. The traditional parameterization of the hazard is
a[1]*exp(a[2]*x) + a[3] + a[4]*exp(a[5]*x) whereas the demohaz
parameterization is b[1]*exp(b[2]*x) + b[3] + b[5]*exp(b[5]*(x-b[4])). They
are related via

b[1] = a[1]
b[2] = a[2]
b[3] = a[3]
b[4] = log(a[5]/a[4])/a[5]
b[5] = a[5]
}
