% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{usher3_nll}
\alias{usher3_nll}
\title{Wrapper function for the negative log-likelihood of the Usher 3 model}
\usage{
usher3_nll(paramVect, ageVect, illVect, x0 = 0)
}
\arguments{
\item{paramVect}{The parameter vector for the Usher illness-death model}

\item{ageVect}{The vector of ages-at-death}

\item{illVect}{The vector of illness indicators}

\item{x0}{Conditional starting age [default: 0]}
}
\value{
The negative log-likelihood value
}
\description{
Wrapper function for the negative log-likelihood of the Usher 3 model
}
