% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{fit_siler}
\alias{fit_siler}
\title{Maximum likelihood fit for the Siler distribution}
\usage{
fit_siler(
  x,
  x0 = 0,
  b0 = c(0.175, 1.4, 0.368 * 0.01, log(0.917 * 0.1/(0.075 * 0.001))/(0.917 * 0.1), 0.917
    * 0.1),
  calc_hessian = FALSE,
  num_cyc = 1000,
  lr = 1e-05,
  miniter = 1000,
  maxiter = 10000,
  verbose = FALSE,
  show_plot = FALSE,
  report_period = 50
)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{x0}{The conditional starting age. The default is 0 (no offset).}

\item{b0}{Initial value for optimization. Default from Gage and Dyke 1986,
Table 2, Level 15. This is also used for the baseline age in the
tranformed negative log-likelihood calculation.}

\item{calc_hessian}{Whether to calculate the Hessian (default FALSE)}

\item{num_cyc}{Number of cycles for tempering}

\item{lr}{The learning rate for gradient descent
(default: 1e-5)}

\item{miniter}{The minimum number of iterations for gradient descent}

\item{maxiter}{The maximum number of iterations for gradient descent}

\item{verbose}{Whether to print out optimization information
(default: FALSE)}

\item{show_plot}{Whether to plot progress (default: FALSE)}

\item{report_period}{How frequently to report information, which includes
printing out information (if verbose is TRUE) and plotting the progress
(if show_plot is TRUE). For tempering, this is with respect to cycles and
for gradient descent this is with respect to iterations.}
}
\value{
A list consisting of the best-fit parameter vector, b, a fit object
  containing detailed information about the fit, and, optionally, the
  Hessian of the negative log-likelihood.
}
\description{
Do a maximum likelihood Siler fit given the input vector of ages, x. An
initial stochastic optimization is done parallel tempering. The best
parameter vector from the tempering is used to initialize a fine tuning
using vanilla gradient descent.
}
