% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{fit_siler}
\alias{fit_siler}
\title{Maximum likelihood fit for the Siler distribution}
\usage{
fit_siler(
  x,
  x0 = 0,
  b0 = c(0.175, 1.4, 0.368 * 0.01, log(0.917 * 0.1/(0.075 * 0.001))/(0.917 * 0.1), 0.917
    * 0.1),
  calc_hessian = FALSE,
  lr = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{x}{Locations at which to evaluate probability density function}

\item{x0}{The conditional starting age. The default is 0 (no offset).}

\item{b0}{Initial value for optimization. Default from Gage and Dyke 1986,
Table 2, Level 15. This is also used for the baseline age in the
tranformed negative log-likelihood calculation.}

\item{calc_hessian}{Whether to calculate the Hessian (default FALSE)}

\item{lr}{The learning rate to use for fine tuning using gradient descent
(default: 1e-5)}

\item{verbose}{Whether to print out optimization information
(default: FALSE)}
}
\value{
A list consisting of the fit (on the transformed variable bbar) and
  maximum likelihood estimate of b. Optionally, the Hessian of the
  log-likelihood is returned, which allows estimation of the standard errors
  of the maximum likelihood estimate via the observed Fisher information
  matrix.
}
\description{
Do a maximum likelihood Siler fit given the input vector of ages, x. To
allow unconstrained optimization to be used, optimization is done on the
transformed variable bbar (see the description for the function
fast_transformed_nllsiler).
}
