% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{hesiler}
\alias{hesiler}
\title{Calculate the Hessian for the Siler distribution}
\usage{
hesiler(a, x, x0 = 0)
}
\arguments{
\item{a}{The parameter vector}

\item{x}{The parameter vector of ages}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The Hessian matrix (5 by 5)
}
\description{
Calculate the Hessian for the Siler distribution
}
