% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{hsiler}
\alias{hsiler}
\title{Calculate the Siler hazard}
\usage{
hsiler(x, b)
}
\arguments{
\item{x}{The vector of ages}

\item{b}{The parameter vector}
}
\value{
The hazard evaluated at the locations in the input vector x
}
\description{
Calculate the Siler hazard at the locations in the input vector x given
the parameter vector a. See details for a definition of the Siler hazard
and related quantities.
}
\details{
The Siler hazard is

\deqn{\lambda(x) = b_1 \exp(-b_2 \, x) + b_3 + b_5 \exp(b_5 \, (x - b_4))}

where \eqn{b_i} (index notation) is the parameter vector. The cumulative
hazard is found by integrating the hazard from some initial point \eqn{x_0}
to \eqn{x},

\deqn{\Lambda(x_0,x) = - \frac{b_1}{b_2} (e^{-b_2 x}
                       - e^{-b_2 x_0}) + b_3 (x-x_0)
                       + (e^{b_5 (x - b_4)} - e^{b_5 (x_0 - b_4)})}.

For all functions, if \eqn{x_0} (optional) is not input it is set equal to
0. The survival and cumulative density function are, respectively,

\deqn{S(x_0,x) = \exp(-\Lambda(x_0,x))
               = \frac{S(0,x)}{S(0,x_0)} = \frac{S(x)}{S(x_0)}}

and

\deqn{F(x_0,x) = 1 - S(x_0,x)}.

The probability density function is

\deqn{f(x_0,x) = \frac{d}{dx} F(x_0,x) = \lambda(x) S(x_0,x)}.

\code{dsiler} gives the probability density function, \code{psiler} the
cumulative distribution function, \code{qsiler} the quantile function,
\code{rsiler} random draws, \code{hsiler} the hazard, \code{chsiler} the
cumulative hazard, \code{ssiler} the survival, \code{nllsiler} the negative
log-likelihood, and \code{hesiler} the Hessian. fit_siler performs a robust,
maximum likelihood fit to yield the best-fit parameter vector a.
}
