% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{load_demohaz_data}
\alias{load_demohaz_data}
\title{Load datasets in the demohaz R package}
\usage{
load_demohaz_data(dataset_name)
}
\arguments{
\item{dataset_name}{Name of the dataset to load}
}
\value{
The requested dataset
}
\description{
Load datasets in the demohaz R package. This is done with a function to
support reformating of the raw data text files and to make the associated
tests clearer. The following datasets can be loaded:
}
\details{
Ache       The Ache lifetable (see demogR::life.table)
Hadza      The Hadza lifetable (see demogR::life.table)
Hiwi       The Hiwi lifetable (see demogR::life.table)
Tsimane    The Tsimane lifetable (see demogR::life.table)
}
