% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{nllsiler}
\alias{nllsiler}
\title{Calculate the negative log-likelihood for the Siler distribution}
\usage{
nllsiler(a, x, x0 = 0)
}
\arguments{
\item{a}{The parameter vector}

\item{x}{The parameter vector of ages}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The vector of negative log-likelihoods (same length as x)
}
\description{
Calculate the negative log-likelihood for the Siler distribution
}
