% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{p_12}
\alias{p_12}
\title{Calculate the transition probability p_12(x0, x)}
\usage{
p_12(x, k1, k2, b_siler, x0 = 0, x_cut = Inf)
}
\arguments{
\item{x}{The age at which to evaluate the transition probability}

\item{k1}{The constant transition rate from the healthy state to the ill
state (before x_cut)}

\item{k2}{The factor by which the mortality hazard out of the ill state is
larger than that out of the healthy state}

\item{b_siler}{The parameter vector for the Siler hazard model (baseline
mortality)}

\item{x0}{The conditional starting age [default: 0]}

\item{x_cut}{The age at which the well-to-ill transition hazard becomes
zero [default: Inf]. When x_cut = Inf, this reduces to the standard
constant hazard model.}
}
\value{
The transition probability p_12(x0, x)
}
\description{
This function computes the probability of being in the ill state at age x,
given that the individual started in the well state at age x0. This accounts
for transitioning from well to ill at some intermediate age y and then
surviving in the ill state to age x.
}
\details{
Computes the probability of transitioning from well to ill between ages x0 
and x. When x0 = 0: p_12(0, x) = k1 * [S_13(0, x)]^k2 * integral from 0 to 
min(x_cut, x) of exp(-k1 * y) * [S_13(0, y)]^(1 - k2) dy

The integral is computed numerically using R's integrate() function.
}
