% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{psiler}
\alias{psiler}
\title{Calculate the Siler cumulative distribution function (CDF) value}
\usage{
psiler(x, a, x0 = 0)
}
\arguments{
\item{x}{The vector of ages}

\item{a}{The parameter vector}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The cumulative distribution function evaluated at the locations in
the input vector x
}
\description{
Calculate the Siler cumulative distribuition function (CDF) at the locations
in the input vector x given the parameter vector a. Optionally, a
conditional starting age can be input. See details in the documentation for
hsiler for a definition of the Siler hazard and related quantities,
including the CDF.
}
