% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler.R
\name{qsiler}
\alias{qsiler}
\title{Calculate the the inverse of the cumulative distribution function (CDF)}
\usage{
qsiler(qvect, b, x0 = 0)
}
\arguments{
\item{qvect}{The vector of quantiles}

\item{b}{The parameter vector}

\item{x0}{The conditional starting age [default: 0].}
}
\value{
The age, x, evaluated at the locations in the input vector qvect
}
\description{
Calculate x given the input quantiles qvect. That is, solve the inverse
problem for the cumulative distribution. This is accomplished using the
following change of variables:

xbar = b5*(x - x0) / (b5*x + 1)

or, equivalently,

x = ((xbar/b5) + x0)/(1-xbar).

xbar lies between 0 to 1, which allows R's root-finder to be invoked on a
bounded interval. Normalizing by b5 regularizes the root finding for large
ages where the positive exponential term dominates the hazard.

See details in the documentation for hsiler for a definition of the Siler
hazard and related quantities.
}
