% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{sample_usher3}
\alias{sample_usher3}
\title{Sample from the Usher 3 distribution using importance sampling}
\usage{
sample_usher3(N, th, dx, xmax, x_mid = NA, infant_prop = NA, area_tol = 1e-06)
}
\arguments{
\item{N}{The number of samples to generate.}

\item{th}{A vector of parameters for the Usher model.}

\item{dx}{The spacing for the discretized sampling.}

\item{xmax}{The maximum age for the discretized sampling.}

\item{x_mid}{The midpoint age for age filtration (optional).}

\item{infant_prop}{The proportion of infants for age filtration (optional).}

\item{area_tol}{The tolerance for checking the area under the density
curves (default: 1e-6).}
}
\value{
A list containing the sampled ages (x), illness indicators (ill),
  and the calculated densities (rho1 and rho2).
}
\description{
This function generates samples from the Usher 3 distribution using
importance sampling. It allows for optional age filtration based on a
specified midpoint age and infant proportion.
}
\examples{
th <- c(2e-2, 1.2, 0.175, 1.40, 0.368, 0.01, 0.075, 0.001, 0.917, 0.1)
sample_usher3(100, th, 0.01, 120)

}
