% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{usher3}
\alias{usher3}
\alias{usher3_rho1}
\alias{usher3_rho2}
\alias{usher3_integrand}
\title{Usher 3 Illness-Death Model Functions}
\usage{
usher3_rho1(x, k1, b_siler, x0 = 0, x_cut = Inf)

usher3_rho2(x, k1, k2, b_siler, x0 = 0, x_cut = Inf)

usher3_integrand(y, k1, k2, b_siler, x_cut = Inf)
}
\arguments{
\item{x}{The vector of ages}

\item{k1}{The transition rate from the healthy state to the ill state}

\item{b_siler}{The parameter vector for the Siler hazard model}

\item{x0}{The conditional starting age [default: 0]}

\item{x_cut}{The age at which the well-to-ill transition hazard becomes
zero [default: Inf]. When x_cut = Inf, this reduces to the standard
constant hazard model.}

\item{k2}{The factor by which the mortality hazard out of the ill state is
larger than that out of the healthy state}

\item{y}{The variable of integration}
}
\value{
The density rho1 evaluated at the locations in the input vector x

The density rho2 evaluated at the locations in the input vector x

The integrand evaluated at y
}
\description{
This file contains functions related to the Usher 3
illness-death model,  including the calculation of the pseudo-densities rho1
and rho2, and the integrand used in the calculation of rho2.
}
\details{
The Usher 3 illness-death model is a multi-state model that
describes the  transitions of individuals among three states: healthy, ill,
and dead. The model is parameterized by transition rates between the states,
and the Siler hazard model is used to describe the baseline mortality.

  The density rho1 represents the probability density of observing an
  individual who died in the healthy state at a given age, while rho2
  represents the probability density of observing an individual who died in
  the ill state at a given age. They are not proper densities since it is
  the some of the two, rho1 + rho2, that integrates to 1; hence, we
  sometimes call them pseudo-densities.

  The calculation of rho2 involves an integral term, which is computed using
  the \code{integrate} function in R. The integrand for this term is given
  by the \code{usher3_integrand} function.
}
