% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{usher3_errors}
\alias{usher3_errors}
\title{Calculate standard errors, z-scores, and p-values for the Usher 3 model}
\usage{
usher3_errors(theta, x, ill, x0 = 0, x_cut = Inf)
}
\arguments{
\item{theta}{The parameter vector for the Usher illness-death model}

\item{x}{The vector of ages-at-death}

\item{ill}{The vector of illness indicators}

\item{x0}{Conditional starting age [default: 0]}

\item{x_cut}{The age at which the well-to-ill transition hazard becomes
zero [default: Inf]}
}
\value{
A data frame with standard errors, z-scores, and p-values
}
\description{
Calculate standard errors, z-scores, and p-values for the Usher 3 model
}
