% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usher3.R
\name{usher3_hessian}
\alias{usher3_hessian}
\title{Calculate the Hessian matrix for the Usher 3 model}
\usage{
usher3_hessian(theta, x, ill, x0 = 0, x_cut = Inf)
}
\arguments{
\item{theta}{The parameter vector for the Usher illness-death model}

\item{x}{The vector of ages-at-death}

\item{ill}{The vector of illness indicators}

\item{x0}{Conditional starting age [default: 0]}

\item{x_cut}{The age at which the well-to-ill transition hazard becomes
zero [default: Inf]}
}
\value{
The Hessian matrix
}
\description{
Calculate the Hessian matrix for the Usher 3 model
}
